Refactor data model settings new field folder
